app/requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.12.13
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==2.0.1
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==38.0.1
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.6
Django==4.0.4
django-cors-headers==3.11.0
django-dbbackup==4.0.1
django-debug-toolbar==3.5.0
django-extensions==3.1.5
django-filter==21.1
django-treebeard==4.5.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
-e git+https://github.com/DMOJ/judge-server.git@bee0ab13d2f35a8fcf34a675de2344e1dca27d2e#egg=dmoj
executing==0.8.3
fake-factory==9999.9.9
Faker==13.11.0
gripcontrol==4.1.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
incremental==22.10.0
ipython==8.2.0
isort==5.11.4
jedi==0.18.1
jsonfield==3.1.0
kombu==5.2.4
lazy-object-proxy==1.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.7.0
natsort==8.1.0
netaddr==0.8.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.6.2
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.7.0
pubcontrol==3.3.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pylint==2.15.9
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyllist==0.3
pylru==1.2.1
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
redis==4.3.3
requests==2.28.0
service-identity==21.1.0
setproctitle==1.2.3
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
termcolor==1.1.0
tomli==2.0.1
tomlkit==0.11.6
traitlets==5.1.1
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.9
uWSGI==2.0.20
vine==5.0.0
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.0
zope.interface==5.5.0
